blind_transfer: Fix race condition when blind transferring party
The blind_transfer REST API test previously initiated the blind transfer
on both channels entering Stasis. It would also place the channels into
the bridge during the same callback, but did not wait for the
ChannelEnteredBridge event. Because of that, at times, the blind transfer
would fail, as one of the channels would not be in the bridge.

This patch triggers the blind transfer when both ChannelEnteredBridge events
have fired.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5477 ad95026d-47f8-4470-b445-30a6b62014f3