Fix the diversion_caller_id test.
The test was actually checking for incorrect identification in its test. The test
had been passing, though, due to the fact that Asterisk was setting incorrect
information in the Diversion header. A recent fix that caused Asterisk to place
corrrect information in the Diversion header caused the test to start failing.
Now the test checks for appropriate information and passes!



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5491 ad95026d-47f8-4470-b445-30a6b62014f3