Fix test failures in dialog-info tests for Asterisk 12.
In Asterisk 12 and Asterisk 13, the way that local and remote URIs
are stored is different. In the testsuite test, this manifests as
Asterisk 12 not including a port and Asterisk 13 including a port.
For the SIPp scenario, I've now made the inclusion of port 5060
optional. If no port is present, then that's equivalent to port 5060
anyway, so now the test will only fail if a port is included and it's
not 5060.

ASTERISK-24243 #close



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5576 ad95026d-47f8-4470-b445-30a6b62014f3